+++ src/servery/listing.py
@@ -791,7 +791,7 @@ _TEMPLATE = """\
 <head>
 <meta charset="utf-8">
 <meta name="viewport" content="width=device-width, initial-scale=1">
-<title>Index of {heading}</title>
+<title>Index of {heading} \N{MIDDLE DOT} servery</title>
 <style>{style}</style>
 </head>
 <body>